split typescript generator into multiple files
